third_party/blink/web_tests/platform/mac-mac14/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/layer_normalization.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] layerNormalization float32 2D tensor axes=[] and options.bias
  assert_array_approx_equals: test layerNormalization float32 property 0, expected 7.862982749938965 +/- 0.0009765625, expected 7.862982749938965 but got 7.859375
[FAIL] layerNormalization float32 4D tensor options.bias and options.axes=[3, 1, 2]
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'build' on 'MLGraphBuilder': Axes must be ordered for layerNormalization."
[FAIL] layerNormalization float32 4D tensor all options
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'build' on 'MLGraphBuilder': Axes must be ordered for layerNormalization."
Harness: the test ran to completion.
